backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-jose[cryptography]==3.3.0
pydantic==2.8.2
pydantic-settings==2.4.0
python-multipart==0.0.9
